+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - he-IL skin language
 - ConvertService.ashx address replaced with converter
 - coauthoring/CommandService.ashx address replaced with command
+- submitForm for all modes
 
 - php-laravel: new integration example
 - php-laravel: upload files to the server